requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-telegram-bot==20.7
python-multipart==0.0.6
sqlalchemy==2.0.23
aiosqlite==0.19.0
aiohttp==3.9.1
python-dotenv==1.0.0